Bike: move general path methods into package.
bike and package move of NiURI.
